[autoplay] created alpha_1
The alpha_1 player picks a random free row, and expands its line of pieces
along that row until it has reached both edges. If the opponent invades that
row, alpha_1 will randomly pick another row or, failing that, revert to an
alpha_0 algorithm.

The alpha_1 player beats an alpha_0 player about 98% of the time, regardless
of if it starts or not.